move tests to test/instrumentation/modules/next/... where it should be
