+++ .eslintrc.json
@@ -19,8 +19,8 @@
     "/lib/opentelemetry-bridge/opentelemetry-core-mini",
     "/test/babel/out.js",
     "/test/lambda/fixtures/esbuild-bundled-handler/hello.js",
-    "/test/nextjs/a-nextjs-app/pages",
-    "/test/nextjs/a-nextjs-app/components",
+    "/test/instrumentation/modules/next/a-nextjs-app/pages",
+    "/test/instrumentation/modules/next/a-nextjs-app/components",
     "/test/sourcemaps/fixtures/lib",
     "/test/sourcemaps/fixtures/src",
     "/test/stacktraces/fixtures/dist",
