+++ package.json
@@ -15,8 +15,8 @@
     "lint:yaml-files": "./dev-utils/lint-yaml-files.sh  # requires node >=10",
     "coverage": "COVERAGE=true ./test/script/run_tests.sh",
     "test": "./test/script/run_tests.sh",
-    "test:deps": "dependency-check start.js index.js 'lib/**/*.js' 'test/**/*.js' '!test/nextjs/a-nextjs-app' --no-dev -i async_hooks -i perf_hooks -i parseurl -i node:http",
-    "test:tav": "tav --quiet && (cd test/nextjs/a-nextjs-app && tav --quiet)",
+    "test:deps": "dependency-check start.js index.js 'lib/**/*.js' 'test/**/*.js' '!test/instrumentation/modules/next/a-nextjs-app' --no-dev -i async_hooks -i perf_hooks -i parseurl -i node:http",
+    "test:tav": "tav --quiet && (cd test/instrumentation/modules/next/a-nextjs-app && tav --quiet)",
     "test:docs": "./test/script/docker/run_docs.sh",
     "test:types": "tsc --project test/types/tsconfig.json && tsc --project test/types/transpile/tsconfig.json && node test/types/transpile/index.js && tsc --project test/types/transpile-default/tsconfig.json && node test/types/transpile-default/index.js",
     "test:babel": "babel test/babel/src.js --out-file test/babel/out.js && cd test/babel && node out.js",
