+++ test/instrumentation/modules/next/a-nextjs-app/.tav.yml
@@ -10,7 +10,7 @@ next-11:
   name: next
   versions: '11.1.0 || 11.1.4'
   node: '>=14.5.0'
-  commands: node next.test.js
+  commands: node ../next.test.js
   peerDependencies:
     - "react@^17.0.2"
     - "react-dom@^17.0.2"
@@ -18,7 +18,7 @@ next:
   name: next
   versions: '12.0.10 || 12.1.6 || 12.2.6 || 12.3.1 || >12.3.1 <13'
   node: '>=14.5.0'
-  commands: node next.test.js
+  commands: node ../next.test.js
   peerDependencies:
     - "react@^18.2.0"
     - "react-dom@^18.2.0"
