+++ test/instrumentation/modules/next/a-nextjs-app/package.json
