+++ test/instrumentation/modules/next/a-nextjs-app/pages/a-throw-in-page-handler.js
