+++ test/instrumentation/modules/next/a-nextjs-app/pages/api/an-api-endpoint-that-throws.js
