+++ test/instrumentation/modules/next/a-nextjs-app/pages/index.js
