+++ test/instrumentation/modules/next/a-nextjs-app/public/elastic-logo.png
