+++ test/instrumentation/modules/next/a-nextjs-app/public/favicon.ico
